# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Social Media API
  description: REST API for Social Media Demo Application
  version: 1.0.0
  contact:
    name: Social Media Team
    email: team@socialmedia.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.socialmedia.example.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /posts:
    get:
      tags:
        - Posts
      summary: Get posts feed
      description: Retrieve a paginated list of posts for the user's feed
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
          description: Page number (0-based)
        - name: size
          in: query
          schema:
            type: integer
            default: 10
          description: Number of posts per page
      responses:
        '200':
          description: Posts retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostList'

    post:
      tags:
        - Posts
      summary: Create a new post
      description: Create a new post with content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  example: "Just deployed my first microservice to Kubernetes!"
                  maxLength: 280
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /posts/{id}:
    get:
      tags:
        - Posts
      summary: Get post by ID
      description: Retrieve a specific post by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Post retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Posts
      summary: Delete a post
      description: Delete a post (only by the author)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '204':
          description: Post deleted successfully
        '403':
          description: Not authorized to delete this post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        displayName:
          type: string
          example: "John Doe"
        avatar:
          type: string
          nullable: true
          example: "https://example.com/avatar.jpg"
        bio:
          type: string
          nullable: true
          example: "Software developer passionate about technology"
        followersCount:
          type: integer
          example: 150
        followingCount:
          type: integer
          example: 75
        postsCount:
          type: integer
          example: 42
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    Post:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        content:
          type: string
          example: "Just deployed my first microservice to Kubernetes!"
        authorId:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        author:
          $ref: '#/components/schemas/User'
        likesCount:
          type: integer
          example: 24
        commentsCount:
          type: integer
          example: 8
        sharesCount:
          type: integer
          example: 3
        isLiked:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    PostList:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 10
        totalElements:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
        first:
          type: boolean
          example: true
        last:
          type: boolean
          example: false
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Invalid input provided"
        status:
          type: integer
          example: 400
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        path:
          type: string
          example: "/api/posts"